fix: PHP 5.6 fatal on every full and incremental sync
`sendFullSync()` and `sendIncrementalSync()` declare `string` and `int`
parameter hints. Those are scalar hints on PHP 7 but plain class hints on
PHP 5.6, where no value can ever satisfy them, so both sync entry points
raise a catchable fatal error on every call. Same root cause as the
`\Throwable` hint in ErrorHandler, on the module's main code path: with
the ErrorHandler fatal fixed, a 5.6 shop that gets past authorization
would have hit this one instead.

`int $limit` was doing real work: the controller feeds it
`Tools::getValue('limit', 50)`, which returns a query string, and the
hint coerced it under PHP 7's non-strict mode. Cast at the controller
boundary instead so the value is an int before it travels. No injection
was possible either way, `DbQuery::limit()` casts to int itself, but the
32 repositories that call `->limit($limit)` uncast now keep receiving an
int as their docblocks claim.

`int $startTime` needed no equivalent, ApiShopContentService sets it from
`time()`.

Types are kept in the docblocks, so PHPStan coverage is unchanged.

+++ controllers/front/apiShopContent.php
@@ -49,8 +49,10 @@ class ps_EventbusApiShopContentModuleFrontController extends ModuleFrontControll
         /** @var string $langIso */
         $langIso = Tools::getValue('lang_iso');
 
+        // cast here rather than through a `int $limit` parameter hint: scalar
+        // hints are class hints on PHP 5.6, which the module still supports
         /** @var int $limit */
-        $limit = Tools::getValue('limit', 50);
+        $limit = (int) Tools::getValue('limit', 50);
 
         /** @var bool $fullSyncRequested */
         $fullSyncRequested = Tools::getValue('full', 0) == 1;
